Fix failing CI build: prevent electron binary download in unit tests
