+++ .gitignore
@@ -40,3 +40,6 @@ tests/**/*.js.map
 
 .claude/worktrees/
 test_output.txt
+
+# User-specific runtime config (may contain secrets)
+AppData/
